\name{ssmsn-package}
\alias{ssmsn-package}
\docType{package}
\title{
Scale-Shape Mixtures of Skew-Normal Distributions
}
\description{
It provides the density and random number generator.
}
\details{
\tabular{ll}{
Package: \tab ssmsn\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2017-01-31\cr
License: \tab GPL (>=2)\cr
}
}

\author{Rocio Maehara \email{rmaeharaa@gmail.com} and Luis Benites \email{lbenitesanchez@gmail.com}
}

\references{
Jamalizadeh, Ahad and Lin, Tsung-I (2016). A general class of scale-shape mixtures of skew-normal distributions: properties and estimation. Computational Statistics, 1-24.
}

\keyword{package}

\seealso{
\code{\link{ssmsn}},
}

\examples{
#See examples for the ssmsn function linked above.
}
